Adjust scales to properly translate Alabama
Previous implementation made it such that the range of the scales always spanned the entirety of the display. This commit makes it such that the scales will make it such that the proportional size of Alabama is still the same based on whichever direction (length or width) is not fully composed of grid squares.